Update configfs structure
Target port names are now GUIDs which get included as the Unit Unique ID
property in the unit directory. Also TPGs are now called tpgt_%d like in
all other target drivers, but we limit it to tpgt_0 only as we don't
support more than one TPGT.

To summarise, the configfs structure is now:
/sys/kernel/config/target/sbp/5254000b8f01e6f6/tpgt_0/lun/lun_0

The GUID should be a valid EUI-64 and follow all the applicable
conventions including the first 3 bytes being the vendor identifier.