mach-bcm2708: Fix UART code in mach/uncompress.h
The UART code was completely broken and caused the RPi to hang at boot
when trying to decompress the kernel. This patch also adds code to set
up the UART at 115200 baud, which is the default speed for the console
later on in the boot process, as it is not set up by the bootloader.

Signed-off-by: Chris Boot <bootc@bootc.net>