spi-bcm27xx: SPI driver for bcm27xx
This is a basic version of an SPI driver for the BCM27xx platform
including BCM2708 as used in the Raspberry Pi.

You'll need to flip the pins into ALT0 mode before you use this driver,
as they are set to a different mode at boot. You can get a tool to do
this at:
http://gromit.cs.man.ac.uk/RaspberryPi/switchPinCtrl.c

Thanks to csoutreach / A Robinson for his driver which I used as an
inspiration. You can find his version here:

http://piface.openlx.org.uk/raspberry-pi-spi-kernel-driver-available-for

IRQ mode is coming next.

Signed-off-by: Chris Boot <bootc@bootc.net>