spi-bcm2708: don't release chipsel when not asked to
We currently assert the appropriate CS line and tear it down again after
each transfer. What we really want to do is keep the CS line active for
all transfers unless we are specifically asked to deassert with struct
spi_transfer->cs_change, or we finish the last transfer in a message.

Signed-off-by: Chris Boot <bootc@bootc.net>