i2c-bcm2708: fix buffer overruns
The driver assumed the hardware wouldn't fill the FIFO with more data
than it was told to process. In the same vein it wrote as much data as
it could until the FIFO was full. Now we check the read/write position
against the message length to make sure we don't read/write too much.

Signed-off-by: Chris Boot <bootc@bootc.net>